pyproject.toml
[tool.poetry]
name = "grpc-python"
version = "0.0.0"
description = "Playground for gRPC with Python"
authors = ["pyto86 <pyto86.me@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.31.0"
grpcio-reflection = "^1.31.0"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.31.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
